requirements.txt
requests==2.31.0
configobj~=5.0.8
pynacl~=1.5.0